Fix author with special chars
In my surname there is the apostrophe character [ -> ' <- ] and the theme was misbehaving (and there were errors in the console).
By escaping the variable the problem goes away.

Maybe it could be better to use quotes to enclose this variable, but at the moment I am not able to test it. This pull request is mostly to report the issue.